+++ doc/whats_new.rst
@@ -144,6 +144,12 @@ Changelog
      By `Olivier Grisel`_.
 
 
+   - Added :class:`linear_model.MultiTaskElasticNetCV` and
+     :class:`linear_model.MultiTaskLassoCV`. In particular,
+     multi-task output, can be cross-validated, without using
+     :class:`linear_model.ElasticNetCV` and :class:`linear_model.LassoCV`
+     per task. By Manoj Kumar.
+
 
 API changes summary
 -------------------